gossip: cleanups and renamings
