+++ src/protocol/gossip/impl/CMakeLists.txt
@@ -5,7 +5,8 @@ libp2p_add_library(p2p_gossip
     common.cpp
     message_builder.cpp
     message_parser.cpp
-    peers.cpp
+    peer_set.cpp
+    peer_context.cpp
     message_cache.cpp
     )
 target_link_libraries(p2p_gossip
